Feat: BaseTableViewCell 생성/적용
