+++ meaning-out/Base/BaseTableViewCell.swift
@@ -1,0 +1,29 @@
+//
+//  BaseTableViewCell.swift
+//  meaning-out
+//
+//  Created by junehee on 7/8/24.
+//
+
+import UIKit
+
+class BaseTableViewCell: UITableViewCell {
+    
+    override init(style: UITableViewCell.CellStyle, reuseIdentifier: String?) {
+        super.init(style: style, reuseIdentifier: reuseIdentifier)
+        configureCellHierarchy()
+        configureCellLayout()
+        configureCellUI()
+    }
+    
+    required init?(coder: NSCoder) {
+        fatalError("init(coder:) has not been implemented")
+    }
+    
+    func configureCellHierarchy() { }
+    
+    func configureCellLayout() { }
+    
+    func configureCellUI() { }
+    
+}
